@@ -1,29 +1,13 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
-    <NoWin32Manifest>true</NoWin32Manifest>
-    <MvcRazorCompileOnPublish>false</MvcRazorCompileOnPublish>
+    <TargetFramework>net6.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.All" Version="2.0.9" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.0.1" />
-    <PackageReference Include="Syncfusion.EJ2.AspNet.Core" Version="*" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <DotNetCliToolReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Tools" Version="2.0.1" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Compile Remove="wwwroot\**\*;node_modules;runtimeconfig.template.json;" />
-    
-    <None Update="wwwroot\**\*">
-      <CopyToPublishDirectory>Always</CopyToPublishDirectory>
-    </None>
-    <None Include="Controllers\**" CopyToOutputDirectory="Always" />
-    <None Include="Views\**" CopyToOutputDirectory="Always" />
+    <PackageReference Include="Syncfusion.EJ2.AspNet.Core" Version="20.4.0.38" />
   </ItemGroup>
 
 </Project>
