@@ -5,7 +5,7 @@ using System.Text;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Html;
 using Microsoft.AspNetCore.Http;
-using Microsoft.Extensions.PlatformAbstractions;
+//using Microsoft.Extensions.PlatformAbstractions;
 using Microsoft.AspNetCore.Hosting;
 using EJ2CoreSampleBrowser.Helpers.BrowserClasses;
 
@@ -23,19 +23,19 @@ namespace EJ2CoreSampleBrowser.Helpers
             get;
             set;
         }
-        public IHostingEnvironment ApplicationPath
+        public IWebHostEnvironment ApplicationPath
         {
             get;
             set;
         }
         private bool SeperateWindow { get; set; }
-        public SourceTabActionResult(string fileName, string seperateWindow, IHostingEnvironment path)
+        public SourceTabActionResult(string fileName, string seperateWindow, IWebHostEnvironment path)
         {
             this.FileName = fileName;
             this.SeperateWindow = seperateWindow == "true";
             this.ApplicationPath = path;
         }
-        public string getContent(IHostingEnvironment path)
+        public string getContent(IWebHostEnvironment path)
         {
             ProductXmlDataEngine xmlEngine = new ProductXmlDataEngine();
             TabType tabType = xmlEngine.GetTabType(this.FileName);
