@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 
-namespace EJ2CoreSampleBrowser.Controllers
+namespace Outlook_Core.Controllers
 {
     public class HomeController : Controller
     {
