@@ -1,4 +1,5 @@
-﻿@model ErrorViewModel
+﻿@using Outlook_Core.Models
+@model ErrorViewModel
 @{
     ViewData["Title"] = "Error";
 }
